loog/requirements.txt
Django==3.2.4
django-crispy-forms==1.11.2
django-filter==2.2.0
djangorestframework==3.12.4
channels==3.0.4
channels-redis==3.3.0
Markdown==3.2.1
mysqlclient==2.0.3
pytz==2021.1
gunicorn==20.0.4
celery==5.1.2
redis==3.5.3
django-redis==5.0.0
nltk==3.6.2
asgiref~=3.4.1
Authlib~=0.15.4
six~=1.16.0
requests==2.25.1
Pillow==7.0.0
django-webpush==0.3.3
